GalaxyNews/requirements.txt
Django
psycopg2-binary
pillow
djangorestframework
djoser
